+++ src/server.ts
@@ -1,6 +1,6 @@
 import app from './app';
-import config from './core/config';
-import logger from './core/logger';
+import config from './config/config';
+import logger from './config/logger';
 
 app.listen(config.PORT, () => {
   logger.info(`Server running on port ${config.PORT}`);
